Do not get all output for list command
List is pretty simply now in that it re-loads all results from the
database and simply prints the (already loaded into memory) output
for the selected result.

This could be painful for large playbook runs, not to mention
unnecessary and inefficient. It was by all means a stop-gap
implementation at the time.

Instead, only load the data that is actually needed for the list
(path and start time for each) using the new PathTime struct and
get the output from the database on demand.